DISCO changes (#43)
* disco_directories

* Listing and extraction from tar file changes, uploaded the final tar file map with the metadata appended to it

* Generic code for listing

* simplified the listing part to only list samples within the metadata from the url

* final changes to listing.js as of now

* changes to downloader.js

* removed python dependency

* removed the python listing file

* Changes made while testing locally

* fixing the troublemaker

* tweaks

* sample name based extraction changes

* Working DISCO integration (tested locally)

* Functional changes to downloader.js

* Made a lookup for the metadata. Renamed the "platform" field to "dataset_technology"

* got rid of the anonymous function

* intial mappings created

* DISCO processing update

* mapping additions to downloader.js

* annotate.sh changes

* rename gene ids to be consistent with other h5ad datasets

* Put in logic to handle single process execution model

* counts for popv compatibility

* [NS-FOREST] add nsforest to build containers script

* added frmatch and updated Uberon mappings with spreadsheet link

---------

+++ scripts/30-annotate.sh
@@ -3,7 +3,10 @@ source constants.sh
 shopt -s extglob
 set -ev
 
-readarray -t DATASET_DIRS < <(node $SRC_DIR/list-downloaded-dirs.js)
+DATASET_DIRS=
+for DIR in $(node $SRC_DIR/list-downloaded-dirs.js); do
+  DATASET_DIRS+=($DIR)
+done
 
 if [[ $RUNNER == "slurm" || $RUNNER == "singularity" ]]; then
   for DIR in ${DATASET_DIRS[@]}; do
@@ -47,15 +50,21 @@ if [[ $RUNNER != "slurm" ]]; then
     for ALGORITHM in azimuth celltypist popv frmatch pan-human-azimuth; do
       if should_run $DIR $ALGORITHM; then
         if [ -e "${DIR}/job-${ALGORITHM}.json" ]; then
-          echo "${SRC_DIR}/run-job.sh ${DIR} ${ALGORITHM}" >> jobs.txt
+          if [ "${MAX_PROCESSES}" == "1" ]; then
+            ${SRC_DIR}/run-job.sh ${DIR} ${ALGORITHM}
+          else
+            echo "${SRC_DIR}/run-job.sh ${DIR} ${ALGORITHM}" >> jobs.txt
+          fi
         fi
       fi
     done
   done
 
-  shuf jobs.txt -o jobs2.txt
-  node src/parallel-jobs.js jobs2.txt
-  rm -f jobs.txt jobs2.txt
+  if [ "${MAX_PROCESSES}" != "1" ]; then
+    shuf jobs.txt -o jobs2.txt
+    node src/parallel-jobs.js jobs2.txt
+    rm -f jobs.txt jobs2.txt
+  fi
 else
   DIRS_FILE="$OUTPUT_DIR/annotate-dirs.txt"
   printf "%s\n" "${DATASET_DIRS[@]}" >$DIRS_FILE
